Add files via upload
A small script I wrote for work to help pick out info from a 'master' excel spreadsheet. I use the python module 'pick' - which gives the option in the terminal with a simple GUI. It works quite well for the basic needs of the script. Next I what to amend it of it can loop and ask if you want to pick again .... rather than straight to 'exit'.